requirements.txt
certifi==2018.4.16
chardet==3.0.4
click==6.7
decorator==4.3.0
Django==1.11.13
django-filter==1.1.0
django-geocoder==0.2.0
djangorestframework==3.8.2
future==0.16.0
geocoder==1.15.1
idna==2.6
Markdown==2.6.11
Pillow==5.1.0
postgis==1.0.4
psycopg2==2.7.4
psycopg2-binary==2.7.4
pytz==2018.4
ratelim==0.1.6
requests==2.18.4
six==1.11.0
urllib3==1.22
